Refactor duplicated PaparazziRule code to common module
