+++ link/src/test/java/com/stripe/android/link/ui/inline/LinkFieldsScreenshotTest.kt
@@ -8,9 +8,9 @@ import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.unit.dp
 import com.stripe.android.link.ui.ErrorMessage
 import com.stripe.android.link.ui.signup.SignUpState
-import com.stripe.android.link.utils.PaparazziRule
-import com.stripe.android.link.utils.screenshots.FontSize
-import com.stripe.android.link.utils.screenshots.SystemAppearance
+import com.stripe.android.screenshottesting.PaparazziRule
+import com.stripe.android.screenshottesting.FontSize
+import com.stripe.android.screenshottesting.SystemAppearance
 import com.stripe.android.uicore.elements.EmailConfig
 import com.stripe.android.uicore.elements.NameConfig
 import com.stripe.android.uicore.elements.PhoneNumberController
