+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/ui/ErrorMessageScreenshotTest.kt
@@ -1,9 +1,9 @@
 package com.stripe.android.paymentsheet.ui
 
-import com.stripe.android.utils.screenshots.FontSize
-import com.stripe.android.utils.screenshots.PaparazziRule
+import com.stripe.android.screenshottesting.FontSize
+import com.stripe.android.screenshottesting.PaparazziRule
 import com.stripe.android.utils.screenshots.PaymentSheetAppearance
-import com.stripe.android.utils.screenshots.SystemAppearance
+import com.stripe.android.screenshottesting.SystemAppearance
 import org.junit.Rule
 import org.junit.Test
