+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/ui/WalletsDividerScreenshotTest.kt
@@ -2,10 +2,10 @@ package com.stripe.android.paymentsheet.ui
 
 import androidx.compose.ui.res.stringResource
 import com.stripe.android.paymentsheet.R
-import com.stripe.android.utils.screenshots.FontSize
-import com.stripe.android.utils.screenshots.PaparazziRule
+import com.stripe.android.screenshottesting.FontSize
+import com.stripe.android.screenshottesting.PaparazziRule
 import com.stripe.android.utils.screenshots.PaymentSheetAppearance
-import com.stripe.android.utils.screenshots.SystemAppearance
+import com.stripe.android.screenshottesting.SystemAppearance
 import org.junit.Rule
 import org.junit.Test
