+++ paymentsheet/src/test/java/com/stripe/android/utils/screenshots/PaymentSheetAppearance.kt
@@ -1,39 +1,9 @@
 package com.stripe.android.utils.screenshots
 
 import android.graphics.Color
-import app.cash.paparazzi.DeviceConfig
-import com.android.resources.NightMode
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.parseAppearance
-
-interface PaparazziConfigOption {
-
-    fun apply(deviceConfig: DeviceConfig): DeviceConfig = deviceConfig
-
-    fun initialize() {
-        // Nothing to do
-    }
-}
-
-enum class SystemAppearance(private val nightMode: NightMode) : PaparazziConfigOption {
-    LightTheme(NightMode.NOTNIGHT),
-    DarkTheme(NightMode.NIGHT);
-
-    override fun apply(deviceConfig: DeviceConfig): DeviceConfig {
-        return deviceConfig.copy(nightMode = nightMode)
-    }
-}
-
-enum class FontSize(val scaleFactor: Float) : PaparazziConfigOption {
-    DefaultFont(scaleFactor = 1f),
-    LargeFont(scaleFactor = 1.5f);
-
-    override fun apply(deviceConfig: DeviceConfig): DeviceConfig {
-        return deviceConfig.copy(
-            fontScale = scaleFactor,
-        )
-    }
-}
+import com.stripe.android.screenshottesting.PaparazziConfigOption
 
 enum class PaymentSheetAppearance(val appearance: PaymentSheet.Appearance) : PaparazziConfigOption {
