+++ screenshot-testing/build.gradle
@@ -1,0 +1,39 @@
+apply from: configs.androidLibrary
+
+dependencies {
+    implementation project(':stripe-ui-core')
+    implementation libs.compose.ui
+    implementation libs.compose.foundation
+    implementation libs.compose.material
+
+    // Implemented here instead of using plugin which implements with testImplementation
+    implementation("app.cash.paparazzi:paparazzi:${versions.paparazzi}") {
+        exclude group: 'org.hamcrest'
+        exclude group: 'org.bouncycastle'
+        exclude group: 'jakarta.activation'
+        exclude group: 'com.android.tools.ddms', module: 'ddmlib'
+        exclude module: 'layoutlib-native-jdk11'
+    }
+}
+
+android {
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion "${versions.androidxComposeCompiler}"
+    }
+
+    packagingOptions {
+        resources {
+            pickFirsts += [
+                    'META-INF/AL2.0',
+                    'META-INF/LGPL2.1',
+                    'META-INF/DEPENDENCIES',
+                    'META-INF/LICENSE.md',
+                    'META-INF/NOTICE.md',
+                    'xsd/catalog.xml'
+            ]
+        }
+    }
+}
