+++ screenshot-testing/src/main/java/com/stripe/android/screenshottesting/PaparazziRule.kt
@@ -1,4 +1,4 @@
-package com.stripe.android.uicore.utils
+package com.stripe.android.screenshottesting
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.PaddingValues
@@ -7,8 +7,10 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalInspectionMode
 import androidx.compose.ui.unit.dp
 import app.cash.paparazzi.DeviceConfig
 import app.cash.paparazzi.Paparazzi
@@ -28,7 +30,6 @@ class PaparazziRule(
 
     private val defaultDeviceConfig = createPaparazziDeviceConfig()
     private val paparazzi = createPaparazzi(defaultDeviceConfig)
-
     private var description: Description? = null
 
     override fun apply(base: Statement, description: Description): Statement {
@@ -66,13 +67,15 @@ class PaparazziRule(
                 }
 
                 paparazzi.snapshot {
-                    StripeTheme {
-                        Surface(color = MaterialTheme.colors.surface) {
-                            Box(
-                                contentAlignment = Alignment.Center,
-                                modifier = boxModifier,
-                            ) {
-                                content()
+                    CompositionLocalProvider(LocalInspectionMode provides true) {
+                        StripeTheme {
+                            Surface(color = MaterialTheme.colors.surface) {
+                                Box(
+                                    contentAlignment = Alignment.Center,
+                                    modifier = boxModifier,
+                                ) {
+                                    content()
+                                }
                             }
                         }
                     }
