+++ settings.gradle
@@ -26,3 +26,4 @@ include ':ml-core:base'
 include ':ml-core:default'
 include ':ml-core:cardscan'
 include ':ml-core:googleplay'
+include ':screenshot-testing'
