+++ stripe-ui-core/src/test/java/com/stripe/android/uicore/elements/DropdownScreenshotTest.kt
@@ -1,7 +1,7 @@
 package com.stripe.android.uicore.elements
 
-import com.stripe.android.uicore.utils.PaparazziRule
-import com.stripe.android.uicore.utils.SystemAppearance
+import com.stripe.android.screenshottesting.PaparazziRule
+import com.stripe.android.screenshottesting.SystemAppearance
 import org.junit.Rule
 import org.junit.Test
