+++ stripe-ui-core/src/test/java/com/stripe/android/uicore/elements/TextFieldUiScreenshotTest.kt
@@ -6,10 +6,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.unit.dp
 import com.stripe.android.core.strings.resolvableString
+import com.stripe.android.screenshottesting.FontSize
 import com.stripe.android.uicore.R
-import com.stripe.android.uicore.utils.FontSize
-import com.stripe.android.uicore.utils.PaparazziRule
-import com.stripe.android.uicore.utils.SystemAppearance
+import com.stripe.android.screenshottesting.PaparazziRule
+import com.stripe.android.screenshottesting.SystemAppearance
 import org.junit.Rule
 import org.junit.Test
